Refactor away trivial classes.
