+++ scenes/levels/pickups/medkit.tscn
@@ -1,42 +1,23 @@
-[gd_scene load_steps=7 format=3 uid="uid://cyk2sakvlse4x"]
+[gd_scene load_steps=5 format=3 uid="uid://cyk2sakvlse4x"]
 
-[ext_resource type="Texture2D" uid="uid://caidr8anqioe7" path="res://assets/spritesheets/powerups-sheet-alpha.png" id="1_6hbu7"]
-[ext_resource type="Script" path="res://scripts/levels/pickups/Medkit.cs" id="1_pjpf3"]
+[ext_resource type="Script" path="res://MechJamIV/Projectile.cs" id="1_pt5qx"]
+[ext_resource type="Texture2D" uid="uid://bldyxfdw0kii2" path="res://assets/spritesheets/spr_sheet_boxes.png" id="2_gneoy"]
 
-[sub_resource type="AtlasTexture" id="AtlasTexture_5cyn3"]
-atlas = ExtResource("1_6hbu7")
-region = Rect2(0, 36, 18, 18)
+[sub_resource type="AtlasTexture" id="AtlasTexture_mb76b"]
+atlas = ExtResource("2_gneoy")
+region = Rect2(0, 0, 32, 32)
 
-[sub_resource type="AtlasTexture" id="AtlasTexture_g1k2c"]
-atlas = ExtResource("1_6hbu7")
-region = Rect2(18, 36, 18, 18)
-
-[sub_resource type="SpriteFrames" id="SpriteFrames_oeigw"]
-animations = [{
-"frames": [{
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_5cyn3")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_g1k2c")
-}],
-"loop": true,
-"name": &"idle",
-"speed": 20.0
-}]
-
-[sub_resource type="CircleShape2D" id="CircleShape2D_7io8x"]
-radius = 9.0
+[sub_resource type="RectangleShape2D" id="RectangleShape2D_l7jv0"]
+size = Vector2(16, 16)
 
 [node name="Medkit" type="RigidBody2D"]
 collision_layer = 0
 collision_mask = 337
-script = ExtResource("1_pjpf3")
+script = ExtResource("1_pt5qx")
 
-[node name="AnimatedSprite2D" type="AnimatedSprite2D" parent="."]
-sprite_frames = SubResource("SpriteFrames_oeigw")
-animation = &"idle"
-autoplay = "idle"
+[node name="Sprite2D" type="Sprite2D" parent="."]
+scale = Vector2(0.5, 0.5)
+texture = SubResource("AtlasTexture_mb76b")
 
 [node name="CollisionShape2D" type="CollisionShape2D" parent="."]
-shape = SubResource("CircleShape2D_7io8x")
+shape = SubResource("RectangleShape2D_l7jv0")
